% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contrlengor.R
\name{c_e}
\alias{c_e}
\title{control de engorda}
\usage{
c_e(df1, d = 30, ds = 7)
}
\arguments{
\item{df(dataframe)a}{nustros datos originales los forzamos a ser un dataframe.}
}
\value{
un data frame con la gdp,conversion alimenticia,costo alimentacion y una grafica de la reacion del tiempo con el peso.
}
\description{
realiza el control de la produccion en un sistema de engorda, con la finanlidad de conocer como se esta desarrollado nuestro sistema.
}
