% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contrlmater.R
\name{c_m}
\alias{c_m}
\title{contro maternidad}
\usage{
c_m(df2)
}
\arguments{
\item{df(dataframe)forzamos}{a nustros datos originales a ser un dataframe}
}
\value{
un dataframe con los porcentajes de mortalidad, promedio de cerdos nacidos,prolificidad y productividad de cerdas.
}
\description{
realiza el control de la maternidad en un sistema de maternidad con la finalidad de conocer el desarrollo de nustro sistema.
}
