% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contrldstt.R
\name{cdl}
\alias{cdl}
\title{control de lechones}
\usage{
cdl(df)
}
\arguments{
\item{df(dataframe)nustros}{datos originales los forzamos a ser un dataframe}
}
\value{
un dataframe con los promedios de peso por animal, total de lechones vivos y muertos, el numero de lechones muertos y la mortalidad predestete.
}
\description{
realiza el control de la produccion dentro de un sistema de crianza de lechones con la finalidad de conocer parametros que nos indiquen el desarrollo de nuestro sistema.
}
